requirements.txt
pyyaml
segno
